fixed reading can pin address

+++ arduino_firmware/arduino_firmware.ino
@@ -1,6 +1,11 @@
 //-----------------------------------------------------------------------------
 // Daisy Driver firmware
 // 2022-12-21 dan@marginallyclever.com
+//
+// be sure to install "STM32 Boards" from Boards Manager.
+// set tools > board to "Generic stm32f4 series"
+// set tools > board part number to "Black F407VE"
+// set tools > usb support to "CDC (generic 'Serial' supercedes U(S)ART)"
 //-----------------------------------------------------------------------------
 #include "config.h"
 
@@ -29,12 +34,12 @@ void readCANAddress() {
   pinMode(PIN_CAN_ADDR4,INPUT);
   pinMode(PIN_CAN_ADDR5,INPUT);
 
-  CANBusAddress = (digitalRead(PIN_CAN_ADDR0) << 0)
-                | (digitalRead(PIN_CAN_ADDR1) << 1)
-                | (digitalRead(PIN_CAN_ADDR2) << 2)
-                | (digitalRead(PIN_CAN_ADDR3) << 3)
-                | (digitalRead(PIN_CAN_ADDR4) << 4)
-                | (digitalRead(PIN_CAN_ADDR5) << 5);
+  CANBusAddress = ((uint8_t)digitalRead(PIN_CAN_ADDR0) << 0)
+                | ((uint8_t)digitalRead(PIN_CAN_ADDR1) << 1)
+                | ((uint8_t)digitalRead(PIN_CAN_ADDR2) << 2)
+                | ((uint8_t)digitalRead(PIN_CAN_ADDR3) << 3)
+                | ((uint8_t)digitalRead(PIN_CAN_ADDR4) << 4)
+                | ((uint8_t)digitalRead(PIN_CAN_ADDR5) << 5);
 }
 
 void printCANPins() {
